Refactor TimeParseResult to TimeParsed, TimeParseResult.RubyTimeParseResult to RubyTimeParsed
